now precision is operational for segmentation
