+++ GANDLF/metrics/__init__.py
@@ -3,8 +3,8 @@ All the metrics are to be called from here
 """
 from GANDLF.losses.regression import MSE_loss, CEL
 from .segmentation import multi_class_dice, hd100, hd95
-from .regression import accuracy, F1_score, classification_accuracy, precision_score, iou_score, balanced_acc_score
-from .generic import recall_score
+from .regression import accuracy, F1_score, classification_accuracy, iou_score, balanced_acc_score
+from .generic import recall_score, precision_score
 
 
 # global defines for the metrics
